Implemented CancelTripAction. Renamed NotificationService -> NotificationCenter's extension
